Restructure component folder
