+++ src/editor/SingleObjectEditor.tsx
