+++ src/editor/header/EditorDropdownDivider.tsx
